+++ samples/CoreSamples/BasicChat-01MEAI/Program.cs
@@ -1,20 +1,17 @@
-﻿using Azure;
-using Azure.AI.Inference;
+﻿using Azure.AI.OpenAI;
 using Microsoft.Extensions.AI;
 using Microsoft.Extensions.Configuration;
+using System.ClientModel;
 using System.Text;
 
-var githubToken = Environment.GetEnvironmentVariable("GITHUB_TOKEN");
-if(string.IsNullOrEmpty(githubToken))
-{
-    var config = new ConfigurationBuilder().AddUserSecrets<Program>().Build();
-    githubToken = config["GITHUB_TOKEN"];
-}
+var config = new ConfigurationBuilder().AddUserSecrets<Program>().Build();
+var endpoint = config["endpoint"];
+var apiKey = new ApiKeyCredential(config["apikey"]);
+var deploymentName = config["deploymentName"] ?? "gpt-4o-mini";
 
-IChatClient client = new ChatCompletionsClient(
-        endpoint: new Uri("https://models.github.ai/inference"),
-        new AzureKeyCredential(githubToken))
-        .AsIChatClient("Phi-4-mini-instruct");
+IChatClient client = new AzureOpenAIClient(new Uri(endpoint), apiKey)
+        .GetChatClient(deploymentName)
+        .AsIChatClient();
 
 // here we're building the prompt
 StringBuilder prompt = new StringBuilder();