+++ samples/CoreSamples/MEAIFunctions/Program.cs
@@ -1,15 +1,13 @@
-﻿using Azure.AI.Inference;
-using Azure;
+﻿using Azure.AI.OpenAI;
 using Microsoft.Extensions.AI;
 using Microsoft.Extensions.Configuration;
+using System.ClientModel;
 using System.ComponentModel;
 
-var githubToken = Environment.GetEnvironmentVariable("GITHUB_TOKEN");
-if (string.IsNullOrEmpty(githubToken))
-{
-    var config = new ConfigurationBuilder().AddUserSecrets<Program>().Build();
-    githubToken = config["GITHUB_TOKEN"];
-}
+var config = new ConfigurationBuilder().AddUserSecrets<Program>().Build();
+var endpoint = config["endpoint"];
+var apiKey = new ApiKeyCredential(config["apikey"]);
+var deploymentName = config["deploymentName"] ?? "gpt-4o-mini";
 
 ChatOptions options = new ChatOptions
 {
@@ -19,10 +17,9 @@ ChatOptions options = new ChatOptions
 };
 
 
-IChatClient client = new ChatCompletionsClient(
-    endpoint: new Uri("https://models.github.ai/inference"),
-    new AzureKeyCredential(githubToken))
-    .AsIChatClient("gpt-4o-mini")
+IChatClient client = new AzureOpenAIClient(new Uri(endpoint), apiKey)
+    .GetChatClient(deploymentName)
+    .AsIChatClient()
     .AsBuilder()
     .UseFunctionInvocation()
     .Build();